requirements.txt
Flask
Flask-Bcrypt
Flask-Login
Flask-SQLAlchemy
Flask-WTF
Jinja2
MarkupSafe
Pillow
SQLAlchemy	
WTForms	
Werkzeug	
bcrypt
cffi
click
itsdangerous
pip
pycparser
setuptools
six